Split collect alignment into 4 buttons
